+++ src/main/java/com/binar/teekmustbe/service/offer/OfferService.java
@@ -3,8 +3,10 @@ package com.binar.teekmustbe.service.offer;
 
 import com.binar.teekmustbe.entitiy.User;
 
+import java.math.BigDecimal;
+
 
 public interface OfferService {
-    void save(long productId, Integer priceNegotiated, User user);
+    void save(long productId, BigDecimal priceNegotiated, User user);
 
 }
