+++ src/main/java/com/binar/teekmustbe/service/offer/OfferServiceImpl.java
@@ -11,6 +11,8 @@ import com.binar.teekmustbe.service.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 public class OfferServiceImpl implements OfferService{
 
@@ -23,7 +25,7 @@ public class OfferServiceImpl implements OfferService{
     @Autowired
     UserService userService;
 
-    public void save(long productId, Integer priceNegotiated, User user) {
+    public void save(long productId, BigDecimal priceNegotiated, User user) {
         var product = productService.findProductById(productId);
         var offer = new Offer();
         offer.setUserId(user);
